MAJOR refactor; separate code and notebooks and scripts for each project
