+++ .gitignore
@@ -7,3 +7,4 @@ logs/
 models/
 result/
 notebooks/debug
+*.egg-info
