+++ envs/acu_env.yaml
@@ -1,0 +1,24 @@
+name: aim2reduce
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - pip=23.3.1
+  - python=3.11.8
+  - pip:
+      - bayesian-optimization==1.4.3
+      - ipython==8.23.0
+      - lightgbm==4.3.0
+      - matplotlib==3.8.4
+      - numpy==1.26.4
+      - pandas==2.2.1
+      - pre-commit==3.7.0
+      - pyarrow==15.0.2
+      - pytest==8.2.1
+      - pytest-xdist==3.6.1
+      - ruff==0.4.3
+      - scikit-learn==1.3.2
+      - scipy==1.11.4
+      - seaborn==0.13.2
+      - tqdm==4.66.2
+      - xgboost==2.0.3
