+++ envs/symp_env.yaml
