+++ pyproject.toml
@@ -7,7 +7,7 @@ name = "preduce"
 description = "Predict undesirable cancer events"
 version = "0.1.0"
 authors = [
-    { name = "Kevin He", email = "Kevin.He@uhn.ca" }
+    { name = "Kevin He", email = "Kevin.He@uhn.ca" },
     { name = "Noke Yuan", email = "Baijiang.Yuan@uhn.ca" }
 ]
 dependencies = [
