+++ scripts/symp/R/bootstrap.R
