+++ src/symp/config.py
