+++ src/symp/viz.py
@@ -15,7 +15,7 @@ import seaborn as sns
 
 import shap
 import pingouin as pg
-from src.util import (
+from .util import (
     calculate_mean_cal_error,
     calculate_max_cal_error,
     find_largest_threshold_interval,
